Renamed test configuration package
