+++ src/test/java/uk/ac/ebi/ampt2d/accession/file/FileAccessioningDatabaseServiceTest.java
@@ -27,7 +27,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import uk.ac.ebi.ampt2d.accession.file.persistence.FileAccessioningDatabaseService;
 import uk.ac.ebi.ampt2d.accession.file.rest.FileDTO;
 import uk.ac.ebi.ampt2d.commons.accession.generators.ModelHashAccession;
-import uk.ac.ebi.ampt2d.test.configurationaccession.FileAccessioningServiceTestConfiguration;
+import uk.ac.ebi.ampt2d.test.configuration.FileAccessioningServiceTestConfiguration;
 
 import java.util.Arrays;
 import java.util.Collection;
