+++ src/test/java/uk/ac/ebi/ampt2d/accession/sample/SampleAccessioningDatabaseServiceTest.java
@@ -26,7 +26,7 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import uk.ac.ebi.ampt2d.accession.sample.persistence.SampleAccessioningDatabaseService;
 import uk.ac.ebi.ampt2d.commons.accession.generators.ModelHashAccession;
-import uk.ac.ebi.ampt2d.test.configurationaccession.SampleAccessioningDatabaseServiceTestConfiguration;
+import uk.ac.ebi.ampt2d.test.configuration.SampleAccessioningDatabaseServiceTestConfiguration;
 
 import java.util.Arrays;
 import java.util.List;
