+++ src/test/java/uk/ac/ebi/ampt2d/test/configuration/SampleAccessioningDatabaseServiceTestConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.test.configurationaccession;
+package uk.ac.ebi.ampt2d.test.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
